move setlist api to other folder
